requirements/dev.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.11.7
asttokens==2.0.5
atomicwrites==1.4.1
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
bandit==1.7.5
beautifulsoup4==4.11.1
black==23.1.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
colorclass==2.2.2
cryptography==37.0.4
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.5
Django==4.0.6
django-admin-rangefilter==0.9.0
django-allauth==0.51.0
django-anymail==8.6
django-appconf==1.0.5
django-braces==1.15.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-extensions==3.2.0
django-imagekit==4.1.0
django-js-asset==2.0.0
django-mailgun-provider==0.2.3
django-mptt==0.13.4
django-nose==1.4.7
django-pytest==0.2.0
django-redis==5.2.0
django-request-logging==0.7.5
django-sendgrid-v5==1.2.1
django-silk==5.0.1
django-summernote==0.8.20.0
django-user-agents==0.4.0
docopt==0.6.2
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
filelock==3.7.1
gitdb==4.0.10
GitPython==3.1.31
gprof2dot==2022.7.29
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.9.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
lazy-object-proxy==1.7.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mdurl==0.1.2
mistune==0.8.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nose==1.3.7
notebook==6.4.12
oauthlib==3.2.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pbr==5.11.1
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
pip-upgrader==1.4.15
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.4
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.14.5
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-ordering==0.6
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
pywin32==304
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
redis==4.5.2
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.2
Send2Trash==1.8.0
sendgrid==6.9.7
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.3.0
starkbank-ecdsa==2.0.3
stevedore==5.0.0
terminado==0.15.0
terminaltables==3.1.10
testpath==0.6.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
traitlets==5.3.0
tzdata==2022.1
ua-parser==0.16.1
urllib3==1.26.11
user-agents==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
wrapt==1.14.1
